feat: add env var for mesos
